Source code/WINDOWS/akamura/ciscomerakiclu/requirements.txt
# Required dependencies to run Cisco Meraki CLU
tabulate
pathlib
datetime
termcolor
requests
rich
setuptools
cryptography
dnspython
ipinfo
scapy
numpy
ipaddress